Sea-orm integration

+++ server/src/misc/responses.rs
@@ -1,7 +1,7 @@
 use async_graphql::SimpleObject;
 use serde::Serialize;
 
-///  This struct should be used in situations where the application
+/// This struct should be used in situations where the application
 /// is not able to return the regular response.
 /// For example:
 /// Lets say the app failed to create a user so it can't send the
@@ -9,7 +9,7 @@ use serde::Serialize;
 /// containing the error message.
 #[derive(Serialize, SimpleObject)]
 pub struct GeneralResponse {
-    pub code: u8,
+    pub code: i32,
     pub message: Option<String>,
     pub error: Option<String>,
 }
